#28 Components refactoring
